fix errors
Passed it through a linter to see how it'd look, and it threw a bunch of errors. I fixed those and it ended up like this. 

Haven't tried running it or anything, but it passes through the linter fine now. I'm curious as to how it ran at all before.